findBreakPoints function
